add ut for hparam
